Installed Microsoft.CodeAnalysis.FxCopAnalyzers

+++ src/MongODM.Core/Repositories/CollectionRepository.cs
@@ -81,7 +81,7 @@ namespace Etherna.MongODM.Repositories
 
             // Get current indexes.
             var currentIndexes = new List<BsonDocument>();
-            using (var indexList = await Collection.Indexes.ListAsync(cancellationToken))
+            using (var indexList = await Collection.Indexes.ListAsync(cancellationToken).ConfigureAwait(false))
                 while (indexList.MoveNext())
                     currentIndexes.AddRange(indexList.Current);
 
@@ -92,11 +92,11 @@ namespace Etherna.MongODM.Repositories
                                      where !newIndexes.Any(newIndex => newIndex.name == indexName)
                                      select index)
             {
-                await Collection.Indexes.DropOneAsync(oldIndex.GetElement("name").Value.ToString(), cancellationToken);
+                await Collection.Indexes.DropOneAsync(oldIndex.GetElement("name").Value.ToString(), cancellationToken).ConfigureAwait(false);
             }
 
             // Build new indexes.
-            await Collection.Indexes.CreateManyAsync(newIndexes.Select(i => i.createIndex), cancellationToken);
+            await Collection.Indexes.CreateManyAsync(newIndexes.Select(i => i.createIndex), cancellationToken).ConfigureAwait(false);
         }
 
         public virtual Task<IAsyncCursor<TProjection>> FindAsync<TProjection>(
@@ -112,8 +112,13 @@ namespace Etherna.MongODM.Repositories
 
         public virtual Task<TResult> QueryElementsAsync<TResult>(
             Func<IMongoQueryable<TModel>, Task<TResult>> query,
-            AggregateOptions? aggregateOptions = null) =>
-            query(Collection.AsQueryable(aggregateOptions));
+            AggregateOptions? aggregateOptions = null)
+        {
+            if (query is null)
+                throw new ArgumentNullException(nameof(query));
+
+            return query(Collection.AsQueryable(aggregateOptions));
+        }
 
         public virtual Task ReplaceAsync(
             object model,
@@ -148,7 +153,7 @@ namespace Etherna.MongODM.Repositories
 
             try
             {
-                return await FindOneAsync(predicate, cancellationToken);
+                return await FindOneAsync(predicate, cancellationToken).ConfigureAwait(false);
             }
             catch (EntityNotFoundException)
             {
@@ -163,10 +168,15 @@ namespace Etherna.MongODM.Repositories
         protected override Task CreateOnDBAsync(TModel model, CancellationToken cancellationToken) =>
             Collection.InsertOneAsync(model, null, cancellationToken);
 
-        protected override Task DeleteOnDBAsync(TModel model, CancellationToken cancellationToken) =>
-            Collection.DeleteOneAsync(
+        protected override Task DeleteOnDBAsync(TModel model, CancellationToken cancellationToken)
+        {
+            if (model is null)
+                throw new ArgumentNullException(nameof(model));
+
+            return Collection.DeleteOneAsync(
                 Builders<TModel>.Filter.Eq(m => m.Id, model.Id),
                 cancellationToken);
+        }
 
         protected override async Task<TModel> FindOneOnDBAsync(TKey id, CancellationToken cancellationToken = default)
         {
@@ -175,7 +185,7 @@ namespace Etherna.MongODM.Repositories
 
             try
             {
-                return await FindOneOnDBAsync(m => m.Id!.Equals(id), cancellationToken: cancellationToken);
+                return await FindOneOnDBAsync(m => m.Id!.Equals(id), cancellationToken: cancellationToken).ConfigureAwait(false);
             }
             catch (EntityNotFoundException)
             {
@@ -193,7 +203,7 @@ namespace Etherna.MongODM.Repositories
 
             var element = await Collection.AsQueryable()
                                           .Where(predicate)
-                                          .SingleOrDefaultAsync(cancellationToken);
+                                          .SingleOrDefaultAsync(cancellationToken).ConfigureAwait(false);
             if (element == default(TModel))
                 throw new EntityNotFoundException("Can't find element");
 
@@ -215,7 +225,7 @@ namespace Etherna.MongODM.Repositories
                 await Collection.ReplaceOneAsync(
                     Builders<TModel>.Filter.Eq(m => m.Id, model.Id),
                     model,
-                    cancellationToken: cancellationToken);
+                    cancellationToken: cancellationToken).ConfigureAwait(false);
             }
             else
             {
@@ -223,7 +233,7 @@ namespace Etherna.MongODM.Repositories
                     session,
                     Builders<TModel>.Filter.Eq(m => m.Id, model.Id),
                     model,
-                    cancellationToken: cancellationToken);
+                    cancellationToken: cancellationToken).ConfigureAwait(false);
             }
 
             // Update dependent documents.
